Run ensemble simulation

Run ensemble simulation

Usage

ensemble(
  sfm,
  n = 10,
  threaded = TRUE,
  qs = c(0.025, 0.975),
  return_sims = TRUE,
  range = NULL,
  cross = FALSE,
  keep_nonnegative_flow = TRUE,
  keep_nonnegative_stock = FALSE,
  keep_unit = TRUE,
  only_stocks = TRUE,
  verbose = TRUE,
  debug = FALSE
)

Arguments

sfm

Stock-and-flow model, object of class sdbuildR_xmile.

n

Number of simulations to run in the ensemble. When range is specified, n defines the number of simulations to run per condition. If each condition only needs to be run once, set n = 1. Defaults to 10.

threaded

If TRUE, run the ensemble in threaded. Defaults to TRUE.

qs

Quantiles to calculate in the summary, e.g. c(0.025, 0.975).

return_sims

If TRUE, return the individual simulations in the ensemble. Set to FALSE to save memory. Defaults to TRUE.

range

List of ranges to vary parameters in the ensemble. Only stocks and constants can be specified. All ranges have to be of the same length if cross = FALSE. Defaults to NULL.

cross

If TRUE, cross the parameters in the range list to generate all possible combinations of parameters. Defaults to TRUE.

keep_nonnegative_flow

If TRUE, keeps original non-negativity setting of flows. Defaults to TRUE.

keep_nonnegative_stock

If TRUE, keeps original non-negativity setting of stocks Defaults to TRUE.

keep_unit

If TRUE, keeps units of variables. Defaults to TRUE.

only_stocks

If TRUE, only save stocks. If FALSE, auxiliaries and flows are saved using a callback function. Only applies if language is set to "Julia" in sim_specs() and no delay functions are used. Defaults to FALSE.

verbose

If TRUE, update on progress. Defaults to FALSE.

debug

If TRUE, print output for debugging. Defaults to FALSE.

Value

Object of class sdbuildR_ensemble, which is a list containing:

success

If TRUE, simulation was successful. If FALSE, simulation failed.

error_message

If success is FALSE, contains the error message.

df

Dataframe with simulation results in long format, if return_sims is TRUE. The iteration number is indicated by column "i". If range was specified, the condition is indicated by column "j".

summary

Dataframe with summary statistics of the ensemble, including quantiles specified in qs. If range was specified, summary statistics are calculated for each condition (j) in the ensemble.

n

Number of simulations run in the ensemble (per condition j if range is specified).

total_n

Total number of simulations run in the ensemble (across all conditions if range is specified).

conditions

Dataframe with the conditions used in the ensemble, if range is specified.

constants

Dataframe with the constant parameters used in the ensemble. The first two columns are "j" (iteration number) and "i" (parameter index), followed by the parameter names.

script

Julia script used for the ensemble simulation.

duration

Duration of the simulation in seconds.

...

Other parameters passed to ensemble