Create ensemble of simulations

Create ensemble of simulations

Usage

ensemble(
  sfm,
  n = 10,
  summary = TRUE,
  qs = c(0.025, 0.975),
  return_sims = FALSE,
  range = NULL,
  keep_nonnegative_flow = TRUE,
  keep_nonnegative_stock = FALSE,
  keep_unit = TRUE,
  only_stocks = FALSE,
  verbose = FALSE,
  debug = FALSE
)

Arguments

sfm

Stock-and-flow model, object of class sdbuildR_xmile.

n

Number of simulations to run in the ensemble. Defaults to 10.

summary

If TRUE, summarize the ensemble simulation results using their mean, variance, median, and quantiles. Defaults to TRUE.

qs

Quantiles to calculate in the summary, e.g. c(0.025, 0.975)

return_sims

If TRUE, return the individual simulations in the ensemble. Defaults to FALSE.

range

List of ranges to vary parameters in the ensemble. If specified, n will be the length of that range. All ranges have to be of the same length. Defaults to NULL.

keep_nonnegative_flow

If TRUE, keeps original non-negativity setting of flows. Defaults to TRUE.

keep_nonnegative_stock

If TRUE, keeps original non-negativity setting of stocks Defaults to TRUE.

keep_unit

If TRUE, keeps units of variables. Defaults to TRUE.

only_stocks

If TRUE, only save stocks. If FALSE, auxiliaries and flows are saved using a callback function. Only applies if language is set to "Julia" in sim_specs() and no delay functions are used. Defaults to FALSE.

verbose

If TRUE, update on progress. Defaults to FALSE.

debug

If TRUE, print output for debugging. Defaults to FALSE.

Value

Object of class sdbuildR_ensemble, which is a list containing:

success

If TRUE, simulation was successful. If FALSE, simulation failed.

filepath_sim

Filepath to the simulation results in .ser format.

script

Julia script used for the ensemble simulation.

duration

Duration of the simulation in seconds.

...

Other parameters passed to ensemble