View all standard units

get_units() yields a dataframe with all standard units in Julia's Unitful package and added custom units by sdbuildR.

Usage

get_units()

Value

Dataframe with units in Julia

See also

Examples

get_units()
#>        description                                                                                                                                                                                                  
#>   [1,] "The meter, the SI base unit of length."                                                                                                                                                                     
#>   [2,] "The second, the SI base unit of time."                                                                                                                                                                      
#>   [3,] "The ampere, the SI base unit of electric current."                                                                                                                                                          
#>   [4,] "The kelvin, the SI base unit of thermodynamic temperature."                                                                                                                                                 
#>   [5,] "The candela, the SI base unit of luminous intensity."                                                                                                                                                       
#>   [6,] "The gram, the SI base unit for weight."                                                                                                                                                                     
#>   [7,] "The mole, the SI base unit for amount of substance."                                                                                                                                                        
#>   [8,] "The steradian, a unit of spherical angle. There are 4pi sr in a sphere."                                                                                                                                    
#>   [9,] "The radian, a unit of angle. There are 2pi rad in a circle."                                                                                                                                                
#>  [10,] "The degree, a unit of angle. There are 360 degrees in a circle."                                                                                                                                            
#>  [11,] "The hertz, an SI unit of frequency, defined as 1 s^-1."                                                                                                                                                     
#>  [12,] "The newton, an SI unit of force, defined as 1 kg * m / s^2."                                                                                                                                                
#>  [13,] "The pascal, an SI unit of pressure, defined as 1 N / m^2."                                                                                                                                                  
#>  [14,] "The joule, an SI unit of energy, defined as 1 N * m."                                                                                                                                                       
#>  [15,] "The watt, an SI unit of power, defined as 1 J / s."                                                                                                                                                         
#>  [16,] "The coulomb, an SI unit of electric charge, defined as 1 A * s."                                                                                                                                            
#>  [17,] "The volt, an SI unit of electric potential, defined as 1 W / A."                                                                                                                                            
#>  [18,] "The ohm, an SI unit of electrical resistance, defined as 1 V / A."                                                                                                                                          
#>  [19,] "The siemens, an SI unit of electrical conductance, defined as 1 Ohm^-1"                                                                                                                                     
#>  [20,] "The farad, an SI unit of electrical capacitance, defined as 1 s^4 * A^2 / (kg * m^2)."                                                                                                                      
#>  [21,] "The henry, an SI unit of electrical inductance, defined as 1 J / A^2."                                                                                                                                      
#>  [22,] "The tesla, an SI unit of magnetic B-field strength, defined as 1 kg / (A * s^2)."                                                                                                                           
#>  [23,] "The weber, an SI unit of magnetic flux, defined as 1 kg * m^2 / (A * s^2)."                                                                                                                                 
#>  [24,] "The lumen, an SI unit of luminous flux, defined as 1 cd * sr."                                                                                                                                              
#>  [25,] "The lux, an SI unit of illuminance, defined as 1 lm / m^2."                                                                                                                                                 
#>  [26,] "The becquerel, an SI unit of radioactivity, defined as 1 nuclear decay per s."                                                                                                                              
#>  [27,] "The gray, an SI unit of ionizing radiation dose, defined as the absorption of 1 J per kg of matter."                                                                                                        
#>  [28,] "The sievert, an SI unit of the biological effect of an ionizing radiation dose."                                                                                                                            
#>  [29,] "The katal, an SI unit of catalytic activity, defined as 1 mol of catalyzed"                                                                                                                                 
#>  [30,] "Percent, a unit meaning parts per hundred. Printed as \"%\"."                                                                                                                                               
#>  [31,] "Permille, a unit meaning parts per thousand. Printed as %"                                                                                                                                                  
#>  [32,] "Permyriad, a unit meaning parts per ten thousand."                                                                                                                                                          
#>  [33,] "Percentmille, a unit meaning parts per hundred thousand."                                                                                                                                                   
#>  [34,] "Permillion, a unit meaning parts per million."                                                                                                                                                              
#>  [35,] "Perbillion, a unit meaning parts per billion (in the short-scale sense), i.e., 10^-9."                                                                                                                      
#>  [36,] "Pertrillion, a unit meaning parts per trillion (in the short-scale sense), i.e., 10^-12."                                                                                                                   
#>  [37,] "Perquadrillion, a unit meaning parts per quadrillion (in the short-scale sense), i.e., 10^-15."                                                                                                             
#>  [38,] "The degree Celsius, an SI unit of temperature, defined such that 0 degrees C = 273.15 K."                                                                                                                   
#>  [39,] "The minute, a unit of time defined as 60 s. The full name `minute` is used instead of the symbol `min` to avoid confusion with the Julia function `min`."                                                   
#>  [40,] "The hour, a unit of time defined as 60 minutes."                                                                                                                                                            
#>  [41,] "The day, a unit of time defined as 24 hr."                                                                                                                                                                  
#>  [42,] "The week, a unit of time, defined as 7 d."                                                                                                                                                                  
#>  [43,] "The year, a unit of time, defined as 365.25 d."                                                                                                                                                             
#>  [44,] "Revolutions per second, a unit of rotational speed, defined as 2pi rad / s."                                                                                                                                
#>  [45,] "Revolutions per minute, a unit of rotational speed, defined as 2pi rad / minute."                                                                                                                           
#>  [46,] "The are, a metric unit of area, defined as 100 m^2."                                                                                                                                                        
#>  [47,] "The hectare, a metric unit of area, defined as 100 a."                                                                                                                                                      
#>  [48,] "The barn, a metric unit of area, defined as 100 fm^2."                                                                                                                                                      
#>  [49,] "The liter, a metric unit of volume, defined as 1000 cm^3."                                                                                                                                                  
#>  [50,] "A unit for measuring molar concentration, equal to 1 mol/L."                                                                                                                                                
#>  [51,] "A quantity equal to the elementary charge, the charge of a single electron, with a value of exactly 1.602,176,634 * 10^-19 C. The letter `q` is used instead of `e` to avoid confusion with Euler's number."
#>  [52,] "The electron-volt, a unit of energy, defined as q*V."                                                                                                                                                       
#>  [53,] "A unit for convenience in angular frequency, equal to 2pi Hz."                                                                                                                                              
#>  [54,] "The bar, a metric unit of pressure, defined as 100 kPa."                                                                                                                                                    
#>  [55,] "The standard atmosphere, a unit of pressure, defined as 101,325 Pa."                                                                                                                                        
#>  [56,] "The torr, a unit of pressure, defined as 1/760 atm."                                                                                                                                                        
#>  [57,] "A quantity representing the speed of light in a vacuum, defined as exactly 2.997,924,58 * 10^8 m/s."                                                                                                        
#>  [58,] "The speed of light in a vacuum, a unit of speed, defined as exactly 2.997,924,58 * 10^8 m/s."                                                                                                               
#>  [59,] "A quantity representing the vacuum permeability constant, defined as 4pi * 10^-7 H / m."                                                                                                                    
#>  [60,] "A quantity representing the vacuum permittivity constant, defined as 1 / (mu0 * c^2)."                                                                                                                      
#>  [61,] "A quantity representing the impedance of free space, a constant defined as mu0 * c."                                                                                                                        
#>  [62,] "A quantity representing the universal gravitational constant, equal to 6.674,30 * 10^-11 m^3 / (kg * s^2) (the CODATA 2018 recommended value)."                                                             
#>  [63,] "A quantity representing the nominal acceleration due to gravity in a vacuum near the surface of the earth, defined by standard to be exactly 9.806,65 m / s^2."                                             
#>  [64,] "A quantity representing Planck's constant, defined as exactly 6.626,070,15 * 10^-34 J * s."                                                                                                                 
#>  [65,] "A quantity representing the reduced Planck constant, defined as h / 2pi."                                                                                                                                   
#>  [66,] "A quantity representing the superconducting magnetic flux quantum, defined as h / (2 * q)."                                                                                                                 
#>  [67,] "A quantity representing the rest mass of an electron, equal to 9.109,383,7015 * 10^-31 kg (the CODATA 2018 recommended value)."                                                                             
#>  [68,] "A quantity representing the rest mass of a neutron, equal to 1.674,927,498,04 * 10^-27 kg (the CODATA 2018 recommended value)."                                                                             
#>  [69,] "A quantity representing the rest mass of a proton, equal to 1.672,621,923,69 * 10^-27 kg (the CODATA 2018 recommended value)."                                                                              
#>  [70,] "A quantity representing the Bohr magneton, equal to q * hbar / (2 * me)."                                                                                                                                   
#>  [71,] "A quantity representing Avogadro's constant, defined as exactly 6.022,140,76 * 10^23 / mol."                                                                                                                
#>  [72,] "A quantity representing the Boltzmann constant, defined as exactly 1.380,649 * 10^-23 J / K."                                                                                                               
#>  [73,] "A quantity representing the molar gas constant, defined as Na * k."                                                                                                                                         
#>  [74,] "A quantity representing the Stefan-Boltzmann constant, defined as pi^2 * k^4 / (60 * hbar^3 * c^2)."                                                                                                        
#>  [75,] "A quantity representing the Rydberg constant, equal to 1.097,373,156,8160 * 10^-7 / m (the CODATA 2018 recommended value)."                                                                                 
#>  [76,] "The unified atomic mass unit, or dalton, a unit of mass defined as 1/12 the mass of an unbound neutral atom of carbon-12, equal to 1.660,539,066,60 * 10^-27 kg (the CODATA 2018 recommended value)."       
#>  [77,] "The nominal acceleration due to gravity in a vacuum near the surface of the earth, a unit of acceleration, defined by standard to be exactly 9.806,65 m / s^2."                                             
#>  [78,] "The gal, a CGS unit of acceleration, defined as 1 cm / s^2."                                                                                                                                                
#>  [79,] "The dyne, a CGS unit of force, defined as 1 g * cm / s^2."                                                                                                                                                  
#>  [80,] "The erg, a CGS unit of energy, defined as 1 dyn * cm."                                                                                                                                                      
#>  [81,] "The barye, a CGS unit of pressure, defined as 1 dyn / cm^2."                                                                                                                                                
#>  [82,] "The poise, a CGS unit of dynamic viscosity, defined as 1 dyn * s / cm^2."                                                                                                                                   
#>  [83,] "The stokes, a CGS unit of kinematic viscosity, defined as 1 cm^2 / s."                                                                                                                                      
#>  [84,] "The gauss, a CGS unit of magnetic B-field strength, defined as 1 Mx / cm^2."                                                                                                                                
#>  [85,] "The oersted, a CGS unit of magnetic H-field strength, defined as 1000 A / (4pi * m)."                                                                                                                       
#>  [86,] "The maxwell, a CGS unit of magnetic flux, defined as 1 Gauss * cm^2."                                                                                                                                       
#>  [87,] "The inch, a US customary unit of length defined as 2.54 cm."                                                                                                                                                
#>  [88,] "The mil, a US customary unit of length defined as 1/1000 inch."                                                                                                                                             
#>  [89,] "The foot, a US customary unit of length defined as 12 inch."                                                                                                                                                
#>  [90,] "The yard, a US customary unit of length defined as 3 ft."                                                                                                                                                   
#>  [91,] "The mile, a US customary unit of length defined as 1760 yd."                                                                                                                                                
#>  [92,] "The angstrom, a metric unit of length defined as 1/10 nm."                                                                                                                                                  
#>  [93,] "The acre, a US customary unit of area defined as 4840 yd^2"                                                                                                                                                 
#>  [94,] "The rankine, a US customary unit of temperature defined as 5/9 K."                                                                                                                                          
#>  [95,] "The degree Fahrenheit, a US customary unit of temperature, defined such that 0 degrees F = 459.67 Ra."                                                                                                      
#>  [96,] "The pound-mass, a US customary unit of mass defined as exactly 0.453,592,37 kg."                                                                                                                            
#>  [97,] "The ounce, a US customary unit of mass defined as 1/16 lb."                                                                                                                                                 
#>  [98,] "The slug, a US customary unit of mass defined as 1 lbf * s^2 / ft."                                                                                                                                         
#>  [99,] "The dram, a US customary unit of mass defined as 1/16 oz."                                                                                                                                                  
#> [100,] "The grain, a US customary unit of mass defined as 1/7000 lb."                                                                                                                                               
#> [101,] "The pound-force, a US customary unit of force defined as 1 lb * ge."                                                                                                                                        
#> [102,] "The calorie, a unit of energy defined as exactly 4.184 J."                                                                                                                                                  
#> [103,] "The British thermal unit, a US customary unit of heat defined by ISO 31-4 as exactly 1055.06 J."                                                                                                            
#> [104,] "Pounds per square inch, a US customary unit of pressure defined as 1 lbf / inch^2."                                                                                                                         
#> [105,] "The common year, a unit of time, defined as 365 d."                                                                                                                                                         
#> [106,] "The common quarter, a unit of time, defined as 365/4 d."                                                                                                                                                    
#> [107,] "The common month, a unit of time, defined as 365/12 d."                                                                                                                                                     
#> [108,] "The quarter, a unit of time, defined as 1/4yr."                                                                                                                                                             
#> [109,] "The month, a unit of time, defined as 1/12yr."                                                                                                                                                              
#> [110,] "The US fluid ounce, to measure liquids"                                                                                                                                                                     
#> [111,] "The US liquid quart, unit of capacity"                                                                                                                                                                      
#> [112,] "The tonne, a metric unit of mass"                                                                                                                                                                           
#> [113,] "The US short ton"                                                                                                                                                                                           
#> [114,] "The US gallon, a unit of volume"                                                                                                                                                                            
#> [115,] "The atom, used in chemistry to quantify microscopic particles"                                                                                                                                              
#> [116,] "The molecule, used in scientific models to count chemical entities"                                                                                                                                         
#> [117,] "The euro, the official currency of the Eurozone"                                                                                                                                                            
#> [118,] "The dollar, the currency of the United States"                                                                                                                                                              
#> [119,] "The pound, the currency of the United Kingdom"                                                                                                                                                              
#>        name                                   
#>   [1,] "m"                                    
#>   [2,] "s"                                    
#>   [3,] "A"                                    
#>   [4,] "K"                                    
#>   [5,] "cd"                                   
#>   [6,] "g"                                    
#>   [7,] "mol"                                  
#>   [8,] "sr"                                   
#>   [9,] "rad"                                  
#>  [10,] "deg"                                  
#>  [11,] "Hz"                                   
#>  [12,] "N"                                    
#>  [13,] "Pa"                                   
#>  [14,] "J"                                    
#>  [15,] "W"                                    
#>  [16,] "C"                                    
#>  [17,] "V"                                    
#>  [18,] "Ohm"                                  
#>  [19,] "S"                                    
#>  [20,] "FALSE"                                
#>  [21,] "H"                                    
#>  [22,] "T"                                    
#>  [23,] "Wb"                                   
#>  [24,] "lm"                                   
#>  [25,] "lx"                                   
#>  [26,] "Bq"                                   
#>  [27,] "Gy"                                   
#>  [28,] "Sv"                                   
#>  [29,] "kat"                                  
#>  [30,] "%"                                    
#>  [31,] "permille"                             
#>  [32,] "pertenthousand"                       
#>  [33,] "pcm"                                  
#>  [34,] "ppm"                                  
#>  [35,] "ppb"                                  
#>  [36,] "ppt"                                  
#>  [37,] "ppq"                                  
#>  [38,] "degC"                                 
#>  [39,] "minute"                               
#>  [40,] "hr"                                   
#>  [41,] "d"                                    
#>  [42,] "wk"                                   
#>  [43,] "yr"                                   
#>  [44,] "rps"                                  
#>  [45,] "rpm"                                  
#>  [46,] "a"                                    
#>  [47,] "ha"                                   
#>  [48,] "b"                                    
#>  [49,] "L"                                    
#>  [50,] "M"                                    
#>  [51,] "q"                                    
#>  [52,] "eV"                                   
#>  [53,] "AngHertz"                             
#>  [54,] "bar"                                  
#>  [55,] "atm"                                  
#>  [56,] "Torr"                                 
#>  [57,] "c0"                                   
#>  [58,] "c"                                    
#>  [59,] "magnetic_constant"                    
#>  [60,] "electric_constant"                    
#>  [61,] "Z0"                                   
#>  [62,] "G"                                    
#>  [63,] "gn"                                   
#>  [64,] "h"                                    
#>  [65,] "reduced_Planck_constant"              
#>  [66,] "superconducting_magnetic_flux_quantum"
#>  [67,] "me"                                   
#>  [68,] "mn"                                   
#>  [69,] "mp"                                   
#>  [70,] "Bohr_magneton"                        
#>  [71,] "Na"                                   
#>  [72,] "k"                                    
#>  [73,] "R"                                    
#>  [74,] "Stefan_Boltzmann_constant"            
#>  [75,] "Rydberg_constant"                     
#>  [76,] "u"                                    
#>  [77,] "ge"                                   
#>  [78,] "Gal"                                  
#>  [79,] "dyn"                                  
#>  [80,] "erg"                                  
#>  [81,] "Ba"                                   
#>  [82,] "P"                                    
#>  [83,] "St"                                   
#>  [84,] "Gauss"                                
#>  [85,] "Oe"                                   
#>  [86,] "Mx"                                   
#>  [87,] "inch"                                 
#>  [88,] "mil"                                  
#>  [89,] "ft"                                   
#>  [90,] "yd"                                   
#>  [91,] "mi"                                   
#>  [92,] "angstrom"                             
#>  [93,] "ac"                                   
#>  [94,] "Ra"                                   
#>  [95,] "degF"                                 
#>  [96,] "lb"                                   
#>  [97,] "oz"                                   
#>  [98,] "slug"                                 
#>  [99,] "dr"                                   
#> [100,] "gr"                                   
#> [101,] "lbf"                                  
#> [102,] "cal"                                  
#> [103,] "btu"                                  
#> [104,] "psi"                                  
#> [105,] "common_yr"                            
#> [106,] "common_quarter"                       
#> [107,] "common_month"                         
#> [108,] "quarter"                              
#> [109,] "month"                                
#> [110,] "fl_oz"                                
#> [111,] "quart"                                
#> [112,] "tonne"                                
#> [113,] "ton"                                  
#> [114,] "US_gal"                               
#> [115,] "atom"                                 
#> [116,] "molecule"                             
#> [117,] "EUR"                                  
#> [118,] "USD"                                  
#> [119,] "GBP"                                  
#>        full_name                               definition                
#>   [1,] "Meter"                                 ""                        
#>   [2,] "Second"                                ""                        
#>   [3,] "Ampere"                                ""                        
#>   [4,] "Kelvin"                                ""                        
#>   [5,] "Candela"                               ""                        
#>   [6,] "Gram"                                  ""                        
#>   [7,] "Mole"                                  ""                        
#>   [8,] "Steradian"                             ""                        
#>   [9,] "Radian"                                ""                        
#>  [10,] "Degree"                                "pi/180"                  
#>  [11,] "Hertz"                                 "1/s"                     
#>  [12,] "Newton"                                "1kg*m/s^2"               
#>  [13,] "Pascal"                                "1N/m^2"                  
#>  [14,] "Joule"                                 "1N*m"                    
#>  [15,] "Watt"                                  "1J/s"                    
#>  [16,] "Coulomb"                               "1A*s"                    
#>  [17,] "Volt"                                  "1W/A"                    
#>  [18,] "Ohm"                                   "1V/A"                    
#>  [19,] "Siemens"                               "1/Ohm"                   
#>  [20,] "Farad"                                 "1s^4*A^2/(kg*m^2)"       
#>  [21,] "Henry"                                 "1J/(A^2)"                
#>  [22,] "Tesla"                                 "1kg/(A*s^2)"             
#>  [23,] "Weber"                                 "1kg*m^2/(A*s^2)"         
#>  [24,] "Lumen"                                 "1cd*sr"                  
#>  [25,] "Lux"                                   "1lm/m^2"                 
#>  [26,] "Becquerel"                             "1/s"                     
#>  [27,] "Gray"                                  "1J/kg"                   
#>  [28,] "Sievert"                               "1J/kg"                   
#>  [29,] "Katal"                                 "1mol/s"                  
#>  [30,] "Percent"                               "1//100"                  
#>  [31,] "Permille"                              "1//1000"                 
#>  [32,] "Pertenthousand"                        "1//10000"                
#>  [33,] "Percentmille"                          "1//100000"               
#>  [34,] "Permillion"                            "1//1000000"              
#>  [35,] "Perbillion"                            "1//1000000000"           
#>  [36,] "Pertrillion"                           "1//1000000000000"        
#>  [37,] "Perquadrillion"                        "1//1000000000000000"     
#>  [38,] "Celsius"                               "(27315//100)K"           
#>  [39,] "Minute"                                "60s"                     
#>  [40,] "Hour"                                  "3600s"                   
#>  [41,] "Day"                                   "86400s"                  
#>  [42,] "Week"                                  "604800s"                 
#>  [43,] "Year"                                  "31557600s"               
#>  [44,] "RevolutionsPerSecond"                  "2pi*rad/s"               
#>  [45,] "RevolutionsPerMinute"                  "2pi*rad/minute"          
#>  [46,] "Are"                                   "100m^2"                  
#>  [47,] "Hectare"                               ""                        
#>  [48,] "Barn"                                  "100fm^2"                 
#>  [49,] "Liter"                                 "m^3//1000"               
#>  [50,] "Molar"                                 "1mol/L"                  
#>  [51,] ""                                      "1.602_176_634e-19*C"     
#>  [52,] "eV"                                    "q*V"                     
#>  [53,] "AngHertz"                              "2pi/s"                   
#>  [54,] "Bar"                                   "100000Pa"                
#>  [55,] "Atmosphere"                            "101325Pa"                
#>  [56,] "Torr"                                  "101325Pa//760"           
#>  [57,] ""                                      "299_792_458*m/s"         
#>  [58,] "SpeedOfLight"                          "1c0"                     
#>  [59,] "magnetic constant"                     "4pi*(1//10)^7*H/m"       
#>  [60,] "electric constant"                     ""                        
#>  [61,] "impedance of free space"               ""                        
#>  [62,] "gravitational constant"                "6.674_30e-11*m^3/kg/s^2" 
#>  [63,] "standard acceleration of gravity"      "9.80665*m/s^2"           
#>  [64,] "Planck constant"                       "6.626_070_15e-34*J*s"    
#>  [65,] "hbar"                                  "h/2pi"                   
#>  [66,] "Superconducting magnetic flux quantum" "h/(2q)"                  
#>  [67,] "electron rest mass"                    "9.109_383_7015e-31*kg"   
#>  [68,] "neutron rest mass"                     "1.674_927_498_04e-27*kg" 
#>  [69,] "proton rest mass"                      "1.672_621_923_69e-27*kg" 
#>  [70,] "Bohr magneton"                         "q*hbar/(2*me)"           
#>  [71,] "Avogadro constant"                     "6.022_140_76e23/mol"     
#>  [72,] "Boltzmann constant"                    "1.380_649e-23*(J/K)"     
#>  [73,] "molar gas constant"                    "Na*k"                    
#>  [74,] "Stefan-Boltzmann constant"             "pi^2*k^4/(60*hbar^3*c^2)"
#>  [75,] "Rydberg constant"                      "10_973_731.568_160/m"    
#>  [76,] "UnifiedAtomicMassUnit"                 "1.660_539_066_60e-27*kg" 
#>  [77,] "EarthGravity"                          "gn"                      
#>  [78,] "Gal"                                   "1cm/s^2"                 
#>  [79,] "Dyne"                                  "1g*cm/s^2"               
#>  [80,] "Erg"                                   "1g*cm^2/s^2"             
#>  [81,] "Barye"                                 "1g/cm/s^2"               
#>  [82,] "Poise"                                 "1g/cm/s"                 
#>  [83,] "Stokes"                                "1cm^2/s"                 
#>  [84,] "Gauss"                                 "(1//10_000)*T"           
#>  [85,] "Oersted"                               "(1_000/4pi)*A/m"         
#>  [86,] "Maxwell"                               "(1//100_000_000)*Wb"     
#>  [87,] "Inch"                                  "(254//10000)*m"          
#>  [88,] "Mil"                                   "(1//1000)*inch"          
#>  [89,] "Foot"                                  "12inch"                  
#>  [90,] "Yard"                                  "3ft"                     
#>  [91,] "Mile"                                  "1760yd"                  
#>  [92,] "Angstrom"                              "(1//10)*nm"              
#>  [93,] "Acre"                                  "(316160658//78125)*m^2"  
#>  [94,] "Rankine"                               "(5//9)*K"                
#>  [95,] "Fahrenheit"                            "(45967//100)Ra"          
#>  [96,] "Pound"                                 "0.45359237kg"            
#>  [97,] "Ounce"                                 "lb//16"                  
#>  [98,] "Slug"                                  "1lb*ge*s^2/ft"           
#>  [99,] "Dram"                                  "oz//16"                  
#> [100,] "Grain"                                 "(32//875)*dr"            
#> [101,] "PoundsForce"                           "1lb*ge"                  
#> [102,] "Calorie"                               "4.184J"                  
#> [103,] "BritishThermalUnit"                    "1055.06J"                
#> [104,] "PoundsPerSquareInch"                   "1lbf/inch^2"             
#> [105,] "Common year"                           "365d"                    
#> [106,] "Common quarter"                        "365/4d"                  
#> [107,] "Common month"                          "365/12d"                 
#> [108,] "Quarter"                               "1/4yr"                   
#> [109,] "Month"                                 "1/12yr"                  
#> [110,] "Fluid ounce"                           "29.5735295625mL"         
#> [111,] "Quart"                                 "946.35cm^3"              
#> [112,] "Tonne"                                 "1000kg"                  
#> [113,] "Ton"                                   "907.18474kg"             
#> [114,] "Gallon"                                "0.003785411784m^3"       
#> [115,] "Atom"                                  "1/6.02214076e23mol"      
#> [116,] "Molecule"                              "1/6.02214076e23mol"      
#> [117,] "Euro"                                  "1"                       
#> [118,] "Dollar"                                "1"                       
#> [119,] "PoundSterling"                         "1"                       
#>        prefix 
#>   [1,] "TRUE" 
#>   [2,] "TRUE" 
#>   [3,] "TRUE" 
#>   [4,] "TRUE" 
#>   [5,] "TRUE" 
#>   [6,] "TRUE" 
#>   [7,] "TRUE" 
#>   [8,] "TRUE" 
#>   [9,] "TRUE" 
#>  [10,] "FALSE"
#>  [11,] "TRUE" 
#>  [12,] "TRUE" 
#>  [13,] "TRUE" 
#>  [14,] "TRUE" 
#>  [15,] "TRUE" 
#>  [16,] "TRUE" 
#>  [17,] "TRUE" 
#>  [18,] "TRUE" 
#>  [19,] "TRUE" 
#>  [20,] "TRUE" 
#>  [21,] "TRUE" 
#>  [22,] "TRUE" 
#>  [23,] "TRUE" 
#>  [24,] "TRUE" 
#>  [25,] "TRUE" 
#>  [26,] "TRUE" 
#>  [27,] "TRUE" 
#>  [28,] "TRUE" 
#>  [29,] "TRUE" 
#>  [30,] "FALSE"
#>  [31,] "FALSE"
#>  [32,] "FALSE"
#>  [33,] "FALSE"
#>  [34,] "FALSE"
#>  [35,] "FALSE"
#>  [36,] "FALSE"
#>  [37,] "FALSE"
#>  [38,] "FALSE"
#>  [39,] "FALSE"
#>  [40,] "FALSE"
#>  [41,] "FALSE"
#>  [42,] "FALSE"
#>  [43,] "TRUE" 
#>  [44,] "FALSE"
#>  [45,] "FALSE"
#>  [46,] "FALSE"
#>  [47,] "FALSE"
#>  [48,] "TRUE" 
#>  [49,] "TRUE" 
#>  [50,] "TRUE" 
#>  [51,] "FALSE"
#>  [52,] "TRUE" 
#>  [53,] "TRUE" 
#>  [54,] "TRUE" 
#>  [55,] "TRUE" 
#>  [56,] "TRUE" 
#>  [57,] "FALSE"
#>  [58,] "FALSE"
#>  [59,] "FALSE"
#>  [60,] "FALSE"
#>  [61,] "FALSE"
#>  [62,] "FALSE"
#>  [63,] "FALSE"
#>  [64,] "FALSE"
#>  [65,] "FALSE"
#>  [66,] "FALSE"
#>  [67,] "FALSE"
#>  [68,] "FALSE"
#>  [69,] "FALSE"
#>  [70,] "FALSE"
#>  [71,] "FALSE"
#>  [72,] "FALSE"
#>  [73,] "FALSE"
#>  [74,] "FALSE"
#>  [75,] "FALSE"
#>  [76,] "FALSE"
#>  [77,] "FALSE"
#>  [78,] "TRUE" 
#>  [79,] "TRUE" 
#>  [80,] "TRUE" 
#>  [81,] "TRUE" 
#>  [82,] "TRUE" 
#>  [83,] "TRUE" 
#>  [84,] "TRUE" 
#>  [85,] "TRUE" 
#>  [86,] "TRUE" 
#>  [87,] "FALSE"
#>  [88,] "FALSE"
#>  [89,] "FALSE"
#>  [90,] "FALSE"
#>  [91,] "FALSE"
#>  [92,] "FALSE"
#>  [93,] "FALSE"
#>  [94,] "FALSE"
#>  [95,] "FALSE"
#>  [96,] "FALSE"
#>  [97,] "FALSE"
#>  [98,] "FALSE"
#>  [99,] "FALSE"
#> [100,] "FALSE"
#> [101,] "FALSE"
#> [102,] "TRUE" 
#> [103,] "FALSE"
#> [104,] "FALSE"
#> [105,] "FALSE"
#> [106,] "FALSE"
#> [107,] "FALSE"
#> [108,] "FALSE"
#> [109,] "FALSE"
#> [110,] "FALSE"
#> [111,] "FALSE"
#> [112,] "FALSE"
#> [113,] "FALSE"
#> [114,] "FALSE"
#> [115,] "FALSE"
#> [116,] "FALSE"
#> [117,] "FALSE"
#> [118,] "FALSE"
#> [119,] "FALSE"