Import Insight Maker model

Import a stock-and-flow model from Insight Maker. Models may be your own or from another user. Importing causal loop diagrams or agent-based models is not supported.

Usage

insightmaker_to_sfm(
  URL,
  filepath_IM,
  keep_nonnegative_flow = TRUE,
  keep_nonnegative_stock = FALSE,
  keep_solver = FALSE
)

Arguments

URL

URL to Insight Maker model. Character.

filepath_IM

File path to Insight Maker model. Only used if URL is not specified. Needs to be a character with suffix .InsightMaker.

keep_nonnegative_flow

If TRUE, keeps original non-negativity setting of flows. Defaults to TRUE.

keep_nonnegative_stock

If TRUE, keeps original non-negativity setting of stocks Defaults to TRUE.

keep_solver

If TRUE, keep the ODE solver as it is. If FALSE, switch to Euler integration in case of non-negative stocks to reproduce the Insight Maker data exactly. Defaults to FALSE.

Value

Stock-and-flow model of class xmile.

Details

Insight Maker models can be imported using either a URL or an Insight Maker file. Ensure the URL refers to a public (not private) model. To download a model file from Insight Maker, first clone the model if it is not your own. Then, go to "Share" (top right), "Export", and "Download Insight Maker file".

See also

Examples

# Load a model from Insight Maker
sfm <- insightmaker_to_sfm(
  URL =
    "https://insightmaker.com/insight/5LxQr0waZGgBcPJcNTC029/Crielaard-et-al-2022"
)
plot(sfm)

# Simulate the model
sim <- simulate(sfm)
plot(sim)