Plot timeseries of ensemble

Visualise ensemble simulation results of a stock-and-flow model. Plot the evolution of stocks over time, with the option of also showing other model variables.

Usage

# S3 method for class 'sdbuildR_ensemble'
plot(
  x,
  i = 1:10,
  j = 1,
  palette = "Dark 2",
  colors = NULL,
  font_family = "Times New Roman",
  font_size = 16,
  ...
)

Arguments

x

Output of ensemble().

i

Indices of the individual trajectories to plot if type = "sims". Defaults to 1:10. Including a high number of trajectories will slow down plotting considerably.

j

Index of the condition to plot. Defaults to 1.

palette

Colour palette. Must be one of hcl.pals().

colors

Vector of colours. If NULL, the colour palette will be used. If specified, will override palette. The number of colours must be equal to the number of variables in the simulation dataframe. Defaults to NULL.

font_family

Font family. Defaults to "Times New Roman".

font_size

Font size. Defaults to 16.

...

Optional parameters

Value

Plot of ensemble simulation.

See also