Set up Julia environment for sdbuildR

Create Julia environment to simulate stock-and-flow models. use_julia() looks for a Julia installation, and will install Julia as well as some required packages if not found. Keep in mind that the first time running use_julia() can take around 5-15 minutes, and all subsequent calls can take around 1 minute.

Usage

use_julia(
  stop = FALSE,
  version = "latest",
  JULIA_HOME = NULL,
  dir = NULL,
  force = FALSE,
  force_install = FALSE,
  ...
)

Arguments

stop

If TRUE, stop active Julia session. Defaults to FALSE.

version

Julia version. Default is "latest", which will install the most recent stable release.

JULIA_HOME

Path to Julia installation. Defaults to NULL to locate Julia automatically.

dir

Directory to install Julia. Defaults to NULL to use default location.

force

If TRUE, force Julia setup to execute again.

force_install

If TRUE, force Julia installation even if existing version is found. Defaults to FALSE.

...

Optional arguments passed to JuliaCall::julia_setup()

Examples

use_julia()